feat: start present mode from detailed view
