+++ frontend/src/stores/slide.js
@@ -1,5 +1,6 @@
 import { ref } from 'vue'
 import { createResource } from 'frappe-ui'
+import { changeSlide } from './slideActions'
 
 const name = ref('')
 const presentation = createResource({
@@ -21,6 +22,19 @@ const slideTransitionDuration = ref(0)
 const position = ref(null)
 const dimensions = ref(null)
 
+const startSlideShow = async () => {
+	await changeSlide(0)
+	const elem = document.querySelector('.slideContainer')
+
+	if (elem.requestFullscreen) {
+		elem.requestFullscreen()
+	} else if (elem.webkitRequestFullscreen) {
+		elem.webkitRequestFullscreen()
+	} else if (elem.msRequestFullscreen) {
+		elem.msRequestFullscreen()
+	}
+}
+
 export {
 	name,
 	presentation,
@@ -34,4 +48,5 @@ export {
 	applyReverseTransition,
 	slideTransition,
 	slideTransitionDuration,
+	startSlideShow,
 }
